area as f32, clippy warnings

+++ src/canvas.rs
@@ -9,7 +9,7 @@ mod structs;
 use structs::Size;
 pub use structs::{Area, Color, CanvasItem, Shape, Text, Image, Font};
 
-use crate::state::{State, Field};
+use crate::state::State;
 
 mod renderer;
 use renderer::Canvas;
@@ -27,7 +27,7 @@ impl CanvasContext {
     pub fn clear(&mut self, color: Color) {
         self.components.clear();
         self.components.push((
-            Area((0, 0), None).into_inner(u16::MAX, &self.size),
+            Area((0.0, 0.0), None).into_inner(u16::MAX, &self.size),
             CanvasItem::Shape(
                 Shape::Rectangle(0.0, self.size.logical()),
                 color
@@ -107,7 +107,7 @@ pub struct CanvasApp<A: CanvasAppTrait> {
 impl<A: CanvasAppTrait> WinitAppTrait for CanvasApp<A> {
     async fn new(window: WinitWindow, scheduler: Scheduler, width: f32, height: f32, scale_factor: f64) -> Self {
         let mut canvas = Canvas::new(window).await;
-        let (width, height) = canvas.resize(width as f32, height as f32);
+        let (width, height) = canvas.resize(width, height);
 
         // #[cfg(any(target_os = "macos", target_os = "ios"))]
         // let path = if let Some(new_path) = get_app_support_path() {
@@ -122,11 +122,11 @@ impl<A: CanvasAppTrait> WinitAppTrait for CanvasApp<A> {
             components: Vec::new(),
             scheduler,
             atlas: CanvasAtlas::default(),
-            size: Size::new(width as f32, height as f32, scale_factor),
+            size: Size::new(width, height, scale_factor),
             state: State::default(),
             triggered_keyboard: Vec::new()
         };
-        let app = A::new(&mut context, width as f32, height as f32).await;
+        let app = A::new(&mut context, width, height).await;
 
         CanvasApp{
             context,
@@ -142,8 +142,8 @@ impl<A: CanvasAppTrait> WinitAppTrait for CanvasApp<A> {
 
     fn on_resize(&mut self, width: f32, height: f32, scale_factor: f64) {
         let (width, height) = self.canvas.resize(width, height);
-        let size = Size::new(width as f32, height as f32, scale_factor);
-        self.app.on_resize(&mut self.context, size.logical().0 as f32, size.logical().1 as f32);
+        let size = Size::new(width, height, scale_factor);
+        self.app.on_resize(&mut self.context, size.logical().0, size.logical().1);
         self.context.size = size;
     }
 
@@ -172,8 +172,8 @@ impl<A: CanvasAppTrait> WinitAppTrait for CanvasApp<A> {
 
     fn on_mouse(&mut self, mut event: MouseEvent) {
         event.position = (
-            self.context.size.scale_logical(event.position.0 as f32),
-            self.context.size.scale_logical(event.position.1 as f32)
+            self.context.size.scale_logical(event.position.0),
+            self.context.size.scale_logical(event.position.1)
         );
         self.app.on_mouse(&mut self.context, event)
     }
